Revert "renames test"
This reverts commit ea09071372fc73fa8ccb33ce39ffb6ce6ac597e8.
