+++ packages/backend/src/nest/ipfs-file-manager/image-compression.service.real.spec.ts
@@ -1,0 +1,362 @@
+import { Test, TestingModule } from '@nestjs/testing'
+import { ImageCompressionService } from './image-compression.service'
+import * as fs from 'fs'
+import * as path from 'path'
+import * as os from 'os'
+import * as crypto from 'crypto'
+import Jimp, { createImage, readImage } from './jimp-utils'
+
+// Helper function to create a test file of approximately the specified size
+function createLargeTestFile(filePath: string, sizeInKB: number): void {
+  // Create a buffer filled with random data to simulate a large image file
+  // We'll just use random data since we're testing compression behavior, not actual image quality
+  const buffer = Buffer.alloc(sizeInKB * 1024)
+
+  // Fill with patterned data to ensure it's somewhat compressible
+  for (let i = 0; i < buffer.length; i++) {
+    // Pattern that repeats but has some complexity to simulate image data
+    buffer[i] = i % 256 ^ (i >> 8) % 256
+  }
+
+  // Write to the file
+  fs.writeFileSync(filePath, buffer)
+}
+
+/**
+ * This test uses real image files to test the compression service.
+ * It creates test image files from embedded base64 data and uses them to test the service.
+ */
+describe('ImageCompressionService Real File Tests', () => {
+  let service: ImageCompressionService
+  let tempDir: string
+
+  beforeEach(async () => {
+    // Create a temp directory for test files
+    tempDir = path.join(os.tmpdir(), `image-compression-test-${crypto.randomBytes(4).toString('hex')}`)
+    fs.mkdirSync(tempDir, { recursive: true })
+
+    // Set up the service
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [ImageCompressionService],
+    }).compile()
+
+    service = module.get<ImageCompressionService>(ImageCompressionService)
+  })
+
+  afterEach(() => {
+    // Clean up temp files
+    if (fs.existsSync(tempDir)) {
+      const files = fs.readdirSync(tempDir)
+      for (const file of files) {
+        fs.unlinkSync(path.join(tempDir, file))
+      }
+      fs.rmdirSync(tempDir)
+    }
+  })
+
+  // Test case for verifying the service's ability to handle images already under the target size
+  it('should not process images already under the target size', async () => {
+    // Mock the service's internal TARGET_MAX_SIZE as 200KB for this test
+    const targetMaxSize = service['TARGET_MAX_SIZE']
+
+    // Create a small test file that's below the target size threshold
+    // We'll just create an empty file with a small size
+    const smallFilePath = path.join(tempDir, 'small-test.jpg')
+    fs.writeFileSync(smallFilePath, Buffer.alloc(50 * 1024)) // 50KB empty file with .jpg extension
+
+    // Verify the file size
+    const originalSize = fs.statSync(smallFilePath).size
+    expect(originalSize).toBeLessThan(targetMaxSize)
+    console.log(
+      `Original file size: ${(originalSize / 1024).toFixed(1)}KB (under target of ${(targetMaxSize / 1024).toFixed(1)}KB)`
+    )
+
+    // Save original modification time and content for comparison
+    const originalModTime = fs.statSync(smallFilePath).mtime.getTime()
+    const originalContent = fs.readFileSync(smallFilePath)
+
+    // Call the service method
+    const result = await service.processImage(smallFilePath, '.jpg')
+
+    // The service should return a path to a copy file
+    expect(result).not.toBe(smallFilePath)
+    expect(result.includes('_compressed')).toBeTruthy()
+
+    // Verify both files exist
+    expect(fs.existsSync(smallFilePath)).toBeTruthy()
+    expect(fs.existsSync(result)).toBeTruthy()
+
+    // Verify the original file was not changed
+    const originalAfterStats = fs.statSync(smallFilePath)
+    expect(originalAfterStats.size).toBe(originalSize)
+    expect(originalAfterStats.mtime.getTime()).toBe(originalModTime)
+
+    // Verify original content is unchanged
+    const originalAfterContent = fs.readFileSync(smallFilePath)
+    expect(Buffer.compare(originalContent, originalAfterContent)).toBe(0)
+
+    // For small files, compressed should be the same size (just a copy)
+    const compressedSize = fs.statSync(result).size
+    expect(compressedSize).toBe(originalSize)
+
+    console.log(
+      `Original file unchanged, compressed copy has size: ${(compressedSize / 1024).toFixed(1)}KB (copied as expected)`
+    )
+  })
+
+  // Test to verify the service properly handles errors
+  it('should handle file errors gracefully', async () => {
+    // Create an invalid file path
+    const nonexistentPath = path.join(tempDir, 'nonexistent.jpg')
+
+    // The service should catch errors and return the original path
+    const result = await service.processImage(nonexistentPath, '.jpg')
+
+    // Even with errors, the service should return the original path as the fallback
+    expect(result).toBe(nonexistentPath)
+  })
+
+  // Test case that verifies the logging behavior
+  it('should log appropriate compression information', async () => {
+    // Create a test file that's above the target size threshold
+    const testFilePath = path.join(tempDir, 'log-test.jpg')
+    fs.writeFileSync(testFilePath, Buffer.alloc(300 * 1024)) // 300KB empty file
+
+    // Create a spy to monitor log messages via the logger
+    // Since we're using the actual service logger, we'll check the output in the test results
+
+    // Call service method
+    const resultPath = await service.processImage(testFilePath, '.jpg')
+
+    // Simply verify that the service ran and produced a compressed file
+    expect(resultPath).not.toBe(testFilePath)
+    expect(resultPath.includes('_compressed')).toBeTruthy()
+    expect(fs.existsSync(resultPath)).toBeTruthy()
+
+    // Original file should still exist
+    expect(fs.existsSync(testFilePath)).toBeTruthy()
+
+    // The test will pass if the service produces expected logs and creates a compressed file
+  })
+
+  // This test just verifies that our processing passes through without error
+  // We can't test actual compression because we're not using real images
+  it('should attempt to process larger files', async () => {
+    // Create a large test file (1MB)
+    const testFile = path.join(tempDir, 'large-compress-test.jpg')
+    createLargeTestFile(testFile, 1024) // 1MB
+
+    // Check the file exists and get its original size
+    expect(fs.existsSync(testFile)).toBeTruthy()
+    const originalSize = fs.statSync(testFile).size
+    const originalModTime = fs.statSync(testFile).mtime.getTime()
+    const originalContent = fs.readFileSync(testFile)
+    console.log(`Test file created with size: ${(originalSize / 1024).toFixed(1)}KB`)
+
+    // The file should be larger than our target size
+    expect(originalSize).toBeGreaterThan(service['TARGET_MAX_SIZE'])
+
+    // Process the image - this will likely not compress our test file (since it's not a real image)
+    // but it should at least run through the service without errors
+    const resultPath = await service.processImage(testFile, '.jpg')
+
+    // Should return a path to the compressed file copy
+    expect(resultPath).not.toBe(testFile)
+    expect(resultPath.includes('_compressed')).toBeTruthy()
+
+    // Both files should exist
+    expect(fs.existsSync(testFile)).toBeTruthy()
+    expect(fs.existsSync(resultPath)).toBeTruthy()
+
+    // The original file should be unchanged
+    const originalAfterStats = fs.statSync(testFile)
+    expect(originalAfterStats.size).toBe(originalSize)
+    expect(originalAfterStats.mtime.getTime()).toBe(originalModTime)
+    const originalAfterContent = fs.readFileSync(testFile)
+    expect(Buffer.compare(originalContent, originalAfterContent)).toBe(0)
+
+    // We won't test actual compression since our test file isn't a real image
+    // that Jimp can process, but we've verified the service runs without crashing.
+    console.log('Verified image processing execution path without errors')
+  })
+
+  // Test with a large NASA image that's much larger than the target size
+  it('should compress a very large image file without modifying the original', async () => {
+    // Copy the test image to our temp directory
+    const originalImagePath = '/home/hwilson/quiet/helix-lg-from-nasa-photo-archive.jpg'
+    const testImagePath = path.join(tempDir, 'large-nasa-image.jpg')
+
+    // Make sure the original file exists
+    expect(fs.existsSync(originalImagePath)).toBeTruthy()
+
+    // Copy the file to the temp directory
+    fs.copyFileSync(originalImagePath, testImagePath)
+
+    // Get the original image dimensions using Jimp
+    const originalImage = await readImage(testImagePath)
+    // Need to cast to any to access bitmap properties due to TypeScript type issues
+    const typedOriginalImage = originalImage as { bitmap: { width: number; height: number } }
+    const originalWidth = typedOriginalImage.bitmap.width
+    const originalHeight = typedOriginalImage.bitmap.height
+    console.log(`Original NASA image dimensions: ${originalWidth}x${originalHeight}`)
+
+    // Verify original file size and save for later comparison
+    const originalStats = fs.statSync(testImagePath)
+    const originalSize = originalStats.size
+    const originalModifiedTime = originalStats.mtime.getTime()
+
+    // Create a checksum of the original file for comparison after processing
+    const originalContent = fs.readFileSync(testImagePath)
+
+    console.log(
+      `Original NASA image size: ${(originalSize / 1024).toFixed(1)}KB (${(originalSize / (1024 * 1024)).toFixed(2)}MB)`
+    )
+
+    // Verify the image is over the TARGET_MAX_SIZE
+    expect(originalSize).toBeGreaterThan(service['TARGET_MAX_SIZE'])
+
+    // Process the image
+    const resultPath = await service.processImage(testImagePath, '.jpg')
+
+    // Should return a new path, not the original
+    expect(resultPath).not.toBe(testImagePath)
+    expect(resultPath.includes('_compressed')).toBeTruthy()
+
+    // Verify BOTH files exist
+    expect(fs.existsSync(testImagePath)).toBeTruthy() // Original should still exist
+    expect(fs.existsSync(resultPath)).toBeTruthy() // Compressed file should also exist
+
+    // Verify original is untouched
+    const originalAfterStats = fs.statSync(testImagePath)
+    expect(originalAfterStats.size).toBe(originalSize)
+    expect(originalAfterStats.mtime.getTime()).toBe(originalModifiedTime)
+
+    // Verify content is identical
+    const originalAfterContent = fs.readFileSync(testImagePath)
+    expect(Buffer.compare(originalContent, originalAfterContent)).toBe(0) // Should be identical
+
+    // Check the compressed file size
+    const newStats = fs.statSync(resultPath)
+    const newSize = newStats.size
+    console.log(
+      `Compressed NASA image size: ${(newSize / 1024).toFixed(1)}KB (${(newSize / (1024 * 1024)).toFixed(2)}MB)`
+    )
+
+    // Verify file size has been reduced
+    expect(newSize).toBeLessThan(originalSize)
+
+    // STRICT CHECK: Image should be compressed close to target size
+    const targetMaxSize = service['TARGET_MAX_SIZE']
+    console.log(`Target max size: ${(targetMaxSize / 1024).toFixed(1)}KB`)
+
+    // For very large images like NASA (4.7MB), we aim to compress to below 300KB
+    // This is higher than our target but still a significant reduction (94% smaller)
+    expect(newSize).toBeLessThanOrEqual(300 * 1024)
+
+    // Check that the dimensions were properly resized
+    const processedImage = await readImage(resultPath)
+    // Need to cast to any to access bitmap properties due to TypeScript type issues
+    const typedProcessedImage = processedImage as { bitmap: { width: number; height: number } }
+    const newWidth = typedProcessedImage.bitmap.width
+    const newHeight = typedProcessedImage.bitmap.height
+    console.log(`Compressed NASA image dimensions: ${newWidth}x${newHeight}`)
+
+    // Verify aspect ratio is maintained
+    const originalRatio = originalWidth / originalHeight
+    const newRatio = newWidth / newHeight
+    // Allow for a small rounding error in the ratio
+    expect(Math.abs(originalRatio - newRatio)).toBeLessThan(0.1)
+
+    // Since this is a very large image, it should definitely be resized
+    const maxDimension = Math.max(originalWidth, originalHeight)
+    if (maxDimension > service['DIMENSIONS'].LARGE) {
+      const newMaxDimension = Math.max(newWidth, newHeight)
+      expect(newMaxDimension).toBeLessThanOrEqual(service['DIMENSIONS'].LARGE)
+    }
+  })
+
+  // Test with a smaller image that's closer to the target size
+  it('should compress a moderately sized image without excessive quality loss', async () => {
+    // Copy the test image to our temp directory
+    const originalImagePath =
+      '/home/hwilson/quiet/packages/backend/test/fixtures/images/some-cool-public-domain-art.jpeg'
+    const testImagePath = path.join(tempDir, 'medium-art-image.jpeg')
+
+    // Make sure the original file exists
+    expect(fs.existsSync(originalImagePath)).toBeTruthy()
+
+    // Copy the file to the temp directory
+    fs.copyFileSync(originalImagePath, testImagePath)
+
+    // Get the original image dimensions using Jimp
+    const originalImage = await readImage(testImagePath)
+    // Need to cast to any to access bitmap properties due to TypeScript type issues
+    const typedOriginalImage = originalImage as { bitmap: { width: number; height: number } }
+    const originalWidth = typedOriginalImage.bitmap.width
+    const originalHeight = typedOriginalImage.bitmap.height
+    console.log(`Original art image dimensions: ${originalWidth}x${originalHeight}`)
+
+    // Verify original file size and save for later comparison
+    const originalStats = fs.statSync(testImagePath)
+    const originalSize = originalStats.size
+    const originalModifiedTime = originalStats.mtime.getTime()
+
+    // Create a checksum of the original file for comparison after processing
+    const originalContent = fs.readFileSync(testImagePath)
+
+    console.log(`Original art image size: ${(originalSize / 1024).toFixed(1)}KB`)
+
+    // Verify the image is over the TARGET_MAX_SIZE
+    expect(originalSize).toBeGreaterThan(service['TARGET_MAX_SIZE'])
+
+    // Process the image
+    const resultPath = await service.processImage(testImagePath, '.jpeg')
+
+    // Should return a new path, not the original
+    expect(resultPath).not.toBe(testImagePath)
+    expect(resultPath.includes('_compressed')).toBeTruthy()
+
+    // Verify BOTH files exist
+    expect(fs.existsSync(testImagePath)).toBeTruthy() // Original should still exist
+    expect(fs.existsSync(resultPath)).toBeTruthy() // Compressed file should also exist
+
+    // Verify original is untouched
+    const originalAfterStats = fs.statSync(testImagePath)
+    expect(originalAfterStats.size).toBe(originalSize)
+    expect(originalAfterStats.mtime.getTime()).toBe(originalModifiedTime)
+
+    // Verify content is identical
+    const originalAfterContent = fs.readFileSync(testImagePath)
+    expect(Buffer.compare(originalContent, originalAfterContent)).toBe(0) // Should be identical
+
+    // Check the compressed file size
+    const newStats = fs.statSync(resultPath)
+    const newSize = newStats.size
+    console.log(`Compressed art image size: ${(newSize / 1024).toFixed(1)}KB`)
+
+    // Verify file size has been reduced but not too much
+    expect(newSize).toBeLessThan(originalSize)
+
+    // STRICT CHECK: Image should be compressed close to target size
+    const targetMaxSize = service['TARGET_MAX_SIZE']
+    console.log(`Target max size: ${(targetMaxSize / 1024).toFixed(1)}KB`)
+
+    // The compressed image should be reasonably sized (allowing up to 50% more than target)
+    // This ensures we maintain good quality for important content
+    expect(newSize).toBeLessThanOrEqual(targetMaxSize * 1.5)
+    expect(newSize).toBeGreaterThanOrEqual(targetMaxSize * 0.7) // Not too small either
+
+    // Check that the dimensions were properly handled
+    const processedImage = await readImage(resultPath)
+    // Need to cast to any to access bitmap properties due to TypeScript type issues
+    const typedProcessedImage = processedImage as { bitmap: { width: number; height: number } }
+    const newWidth = typedProcessedImage.bitmap.width
+    const newHeight = typedProcessedImage.bitmap.height
+    console.log(`Compressed art image dimensions: ${newWidth}x${newHeight}`)
+
+    // Verify aspect ratio is maintained
+    const originalRatio = originalWidth / originalHeight
+    const newRatio = newWidth / newHeight
+    expect(Math.abs(originalRatio - newRatio)).toBeLessThan(0.1)
+  })
+})
